feat: added weight accounting to the leaderboard and loserboard, points are now real

+++ Mr.Fish/Models/FishCatch.cs
@@ -6,7 +6,7 @@ public class FishCatch
     public string FishName { get; init; } = "";
     public string AdjectiveName { get; init; } = "";
     public decimal WeightKg { get; init; }
-    public int Points { get; init; }
+    public decimal Points { get; init; }
     public int Rarity { get; init; }
     public bool IsSpecial { get; init; }
     public DateTime CaughtAt { get; init; } = DateTime.UtcNow;
